Clean up and de-duplicate window keyMaps
